Implemented Scoping:
Modified environment to be a stack, each scope/environment is pushed onto the stack. The current scope is the environment on the top of the stack.

Used this idea to implement lexical scoping.
Essentially, going from the end of the list to the beginning is analogous ot starting at the innermost scope and moving outwards.
This idea is used when looking for variables within scopes.
A new environment is pushed onto the stack when we enter a new scope, and it is popped off the stack when we exit.